feat: tanksDataContext for transferring data between new wine pages + select pe de cuba page
