+++ app/(newWine)/_layout.tsx
@@ -3,12 +3,26 @@ import CustomStatusBar from "@/components/CustomStatusBar";
 import FormFooter from "@/components/FormFooter";
 import SafeAreaView from "@/components/SafeAreaView";
 import { Stack, usePathname } from "expo-router";
-import React, { useState } from "react";
+import React, { useEffect, useState } from "react";
 
 export default function NewWineLayout() {
-  const [nextHref, setNextHref] = useState("");
+  const [nextHref, setNextHref] = useState<string>("");
   const currRoute = usePathname();
-  // const routes = [{array de rotas}];
+  const routes = [
+    {
+      route: "/",
+      nextHref: "/(newWine)/selectPeDeCuba",
+    },
+    {
+      route: "/(newWine)/selectPeDeCuba",
+      nextHref: "/(newWine)/",
+    },
+  ];
+
+  useEffect(() => {
+    const r = routes.find((r) => r.route === currRoute);
+    r ? setNextHref(r.nextHref) : null;
+  }, [currRoute]);
 
   return (
     <>
@@ -24,7 +38,7 @@ export default function NewWineLayout() {
         <FormFooter
           nextHref={nextHref}
           isReturnButtonEnabled={currRoute === "/" ? false : true}
-          // handleDataSubmit={handleDataSubmit}
+          // handleDataSubmit={}
         />
       </SafeAreaView>
     </>
