+++ app/(newWine)/index.tsx
@@ -1,29 +1,17 @@
 import apiInstance from "@/api/apiInstance";
 import SafeAreaView from "@/components/SafeAreaView";
+import TankCard from "@/components/TankCard";
+import { useTanksStore } from "@/store/TanksContext";
+import ITankData from "@/types/ITankData";
 import { useFocusEffect } from "expo-router";
 import * as SecureStore from "expo-secure-store";
 import { Search } from "lucide-react-native";
 import React, { useCallback, useState } from "react";
-import {
-  FlatList,
-  Text,
-  TextInput,
-  TouchableOpacity,
-  View,
-} from "react-native";
-
-interface ITankData {
-  conteudo: string;
-  idConteudo: number;
-  densidade: number;
-  temperatura: number;
-  pressao: number;
-  idDeposito: number;
-  deposito: string;
-}
+import { FlatList, Text, TextInput, View } from "react-native";
 
 export default function SelectMostroView() {
   const [data, setData] = useState<ITankData[]>([]);
+  const { setTanksData } = useTanksStore();
   useFocusEffect(
     // calls the api everytime the screen gets displayed
     useCallback(() => {
@@ -44,7 +32,7 @@ export default function SelectMostroView() {
         }
       );
       setData(response.data.filter((t: ITankData) => t.conteudo === "Mostro"));
-      console.log(response.data);
+      setTanksData(response.data);
     } catch (error) {
       console.error("Erro ao buscar depósitos:", error);
     }
@@ -75,7 +63,7 @@ export default function SelectMostroView() {
             keyExtractor={(item) => item.deposito}
             renderItem={({ item }) => {
               return item.temperatura ? (
-                <Card
+                <TankCard
                   depositId={item.idDeposito}
                   title={item.deposito}
                   isAvailable={"Edge"}
@@ -86,7 +74,7 @@ export default function SelectMostroView() {
                   pressure={item.pressao ? item.pressao : null}
                 />
               ) : (
-                <Card
+                <TankCard
                   depositId={item.idDeposito}
                   title={item.deposito}
                   isAvailable={"Edge"}
@@ -99,85 +87,5 @@ export default function SelectMostroView() {
         </View>
       </SafeAreaView>
     </>
-  );
-}
-
-interface CardProps {
-  depositId: number;
-  title: string;
-  isAvailable: boolean | string;
-  density?: number;
-  temperature?: number;
-  pressure?: number | null;
-  content?: string;
-  contentId?: number;
-}
-
-function Card({
-  title,
-  isAvailable,
-  density = 0,
-  temperature = 0,
-  pressure = 0,
-  depositId,
-  content = "",
-  contentId = 0,
-}: CardProps) {
-  const [isSelected, setIsSelected] = useState(false);
-  return (
-    <View style={{ marginBottom: 16, width: "100%" }}>
-      <View className="bg-white rounded-lg shadow flex-col border border-neutral-250">
-        <View className="flex-row p-4 justify-between items-center">
-          <Text className="text-2xl font-bold">{title}</Text>
-          <TouchableOpacity
-            onPress={() => setIsSelected(!isSelected)}
-            className={
-              "border rounded-md px-2 py-1 border-blue-500" +
-              (isSelected ? " bg-blue-500" : "")
-            }
-          >
-            <Text className={isSelected ? "text-white" : "text-blue-500"}>
-              {isSelected ? "Selecionado" : "Selecionar"}
-            </Text>
-          </TouchableOpacity>
-        </View>
-        {isAvailable === "Edge" && temperature ? (
-          <>
-            <View className="w-full h-[1px] bg-neutral-250"></View>
-            <View className="p-4">
-              <View className="flex-row justify-between items-center">
-                <Text className="text-xl font-light">Densidade:</Text>
-                <View className="flex-row justify-center items-end">
-                  <Text className="text-2xl font-semibold">{density} </Text>
-                  <Text className="text-base font-normal text-neutral-400">
-                    kg/m³
-                  </Text>
-                </View>
-              </View>
-              <View className="flex-row justify-between">
-                <Text className="text-xl font-light">Temperatura:</Text>
-                <View className="flex-row justify-center items-end">
-                  <Text className="text-2xl font-semibold">{temperature} </Text>
-                  <Text className="text-base font-normal text-neutral-400">
-                    °C
-                  </Text>
-                </View>
-              </View>
-              {pressure && (
-                <View className="flex-row justify-between">
-                  <Text className="text-xl font-light">Pressão:</Text>
-                  <View className="flex-row justify-center items-end">
-                    <Text className="text-2xl font-semibold">{pressure} </Text>
-                    <Text className="text-base font-normal text-neutral-400">
-                      Pa
-                    </Text>
-                  </View>
-                </View>
-              )}
-            </View>
-          </>
-        ) : null}
-      </View>
-    </View>
   );
 }
