+++ types/ITankData.ts
@@ -1,0 +1,9 @@
+export default interface ITankData {
+  conteudo: string;
+  idConteudo: number;
+  densidade: number;
+  temperatura: number;
+  pressao: number;
+  idDeposito: number;
+  deposito: string;
+}
